Revert "STAGE-553 dropdown input and options split"
