+++ app/components/basic/Dropdown.js
@@ -2,18 +2,19 @@
  * Created by pposel on 23/01/2017.
  */
 
-import React, { Children, cloneElement } from 'react';
+import React, { Component, PropTypes, Children, cloneElement } from 'react';
 import { Dropdown as DropdownSemanticUiReact, DropdownMenu, DropdownHeader } from 'semantic-ui-react'
-import { createShorthand } from 'semantic-ui-react/dist/commonjs/lib';
+import { createShorthand, useKeyOnly, useKeyOrValueAndKey, isBrowser } from '../../../node_modules/semantic-ui-react/dist/commonjs/lib';
+import Portal from '../../../node_modules/semantic-ui-react/dist/commonjs/addons/Portal';
 
 /**
  * Dropdown is a component which extends [Dropdown](https://react.semantic-ui.com/modules/dropdown) component from Semantic-UI-React framework
+ * and wraps it with use of [Portal](https://react.semantic-ui.com/addons/portal) component from Semantic-UI-React framework.
+ *
  * See [Dropdown](https://react.semantic-ui.com/modules/dropdown) component from Semantic-UI-React framework for details about props and usage details.
  *
  * ## Access
  * `Stage.Basic.Form.Dropdown` or `Stage.Basic.Dropdown`
- *
- * Note: This component is closely related to `Stage.Basic.DropdownPortal` component, which additionally wraps the Dropdown in a Portal from Semantic-UI-React.
  *
  * ## Usage
  * ```
@@ -33,6 +34,39 @@ import { createShorthand } from 'semantic-ui-react/dist/commonjs/lib';
  */
 export default class Dropdown extends DropdownSemanticUiReact {
 
+    componentWillMount() {
+        super.componentWillMount();
+        window.addEventListener('scroll', this.hideOnScroll)
+    }
+
+    componentWillUnmount() {
+        super.componentWillUnmount();
+        window.removeEventListener('scroll', this.hideOnScroll);
+    }
+
+    hideOnScroll = (e) => {
+        if (this.state.open) {
+            this.close(e);
+        }
+    }
+
+    computePopupStyle() {
+        var coords = this.ref.getBoundingClientRect();
+
+        const style = { position: 'absolute' };
+
+        // Do not access window/document when server side rendering
+        if (!isBrowser) return style;
+
+        const { pageYOffset, pageXOffset } = window;
+
+        style.left = Math.round(coords.left + pageXOffset) + 'px';
+        style.top = Math.round(coords.top + coords.height + pageYOffset - 2) + 'px';
+        style.width = coords.width + 'px';
+
+        return style
+    }
+
     renderMenu = () => {
         const { open } = this.state;
         const menuClasses = open ? 'visible' : '';
@@ -40,9 +74,55 @@ export default class Dropdown extends DropdownSemanticUiReact {
 
         const {
             children,
-            header
+            header,
+            basic,
+            button,
+            className,
+            compact,
+            floating,
+            inline,
+            item,
+            labeled,
+            multiple,
+            pointing,
+            search,
+            selection,
+            simple,
+            loading,
+            error,
+            disabled,
+            scrolling
         } = this.props;
 
+        // Classes
+        const classes = [
+            'ui',
+            useKeyOnly(open, 'active visible'),
+            useKeyOnly(disabled, 'disabled'),
+            useKeyOnly(error, 'error'),
+            useKeyOnly(loading, 'loading'),
+            useKeyOnly(basic, 'basic'),
+            useKeyOnly(button, 'button'),
+            useKeyOnly(compact, 'compact'),
+            useKeyOnly(floating, 'floating'),
+            useKeyOnly(inline, 'inline'),
+            useKeyOnly(labeled, 'labeled'),
+            useKeyOnly(item, 'item'),
+            useKeyOnly(multiple, 'multiple'),
+            useKeyOnly(search, 'search'),
+            useKeyOnly(selection, 'selection'),
+            useKeyOnly(simple, 'simple'),
+            useKeyOnly(scrolling, 'scrolling'),
+            useKeyOrValueAndKey(pointing, 'pointing'),
+            'dropdown',
+            'dropdownPortal',
+            className
+        ].filter(e => !_.isEmpty(e)).join(' ');
+
+        var style = {};
+        if (this.ref && open) {
+            style = this.computePopupStyle();
+        }
 
         // single menu child
         // Used only for dropdown menu, we are currently using Popup Menu so I do not port it to the body.
@@ -54,10 +134,12 @@ export default class Dropdown extends DropdownSemanticUiReact {
         }
 
         return (
-            <DropdownMenu {...ariaOptions} className={`${menuClasses} ${name}`} >
-                {createShorthand(DropdownHeader, val => ({content: val}), header)}
-                {this.renderOptions()}
-            </DropdownMenu>
+            <Portal open={open} className={classes}>
+                <DropdownMenu {...ariaOptions} className={`${menuClasses} ${name}`} style={style}>
+                    {createShorthand(DropdownHeader, val => ({ content: val }), header)}
+                    {this.renderOptions()}
+                </DropdownMenu>
+            </Portal>
         )
     }
