+++ app/styles/style.scss
@@ -98,10 +98,6 @@ a:hover {
   overflow-y: initial;
 }
 
-.grid-stack-item-content.auto-z-index.grid-stack-override.widget-content {
-  z-index: auto !important;
-}
-
 .gridStackBottomSpace {
   height: 150px;
 }
@@ -654,6 +650,7 @@ code.hljs {
 
 .ui.modal:not(.overlay) > .content {
   max-height: 480px;
+  overflow: auto;
 }
 
 .ui.modal.unlimited > .content {
